[getIssuerEntity] move getIssuerEntity to shared hooks dir
