+++ src/services/api/userCredentialRequests/userCredentialRequests.hooks.ts
@@ -1,34 +1,12 @@
-import { Hook } from '@feathersjs/feathers';
-
-import logger from '../../../logger';
-import { IssuerEntity } from '../../../entities/Issuer';
 import { handleUserDidAssociation } from '../../hooks/handleUserDidAssociation';
 import { validateCredentialRequest } from '../../hooks/validateCredentialRequest';
-
-export const getDefaultIssuerEntity: Hook = async (ctx) => {
-  const issuerDataService = ctx.app.service('issuerData');
-  let defaultIssuerEntity: IssuerEntity;
-  try {
-    defaultIssuerEntity = await issuerDataService.getDefaultIssuerEntity();
-
-    return {
-      ...ctx,
-      params: {
-        ...ctx.params,
-        defaultIssuerEntity
-      }
-
-    };
-  } catch (e) {
-    logger.error('getDefaultIssuerEntity hook caught an error thrown by issuerDataService.getDefaultIssuerEntity', e);
-    throw e;
-  }
-};
+import { config } from '../../../config';
+import { getIssuerEntity } from '../../hooks/getIssuerEntity';
 
 export const hooks = {
   before: {
     all: [validateCredentialRequest],
-    create: [getDefaultIssuerEntity, handleUserDidAssociation]
+    create: [getIssuerEntity(config.DEFAULT_ISSUER_DID), handleUserDidAssociation]
   },
   after: {}
 };
